test: Merge unit tests for upwind coupling.
